Documentation: Use more domain directives from sphinx
PySide's class pages were documents with the class name as anchor and
a headline with the class name. The content was written out flat
with fully qualified function/property/enum names.

Besides having the fully qualified names in the text, this causes
problems like duplicated module prefixes in the sphinx inventory and
in tooltips.

Change this to
- Use the :class: domain directive from sphinx along with the module
  for classes at the top
- Indent the following class description accordingly
- Change constructors to be listed as __init__() methods
- List enums as nested classes
- Write enums before properties
- Use standard references for methods in the functions TOC
  (which strips the parameters, hence build unique lists).

Task-number: PYSIDE-1106
Task-number: PYSIDE-2215
Change-Id: I92ab39af1c35317cddf47cebbf97da5dc852eb33
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>