build: update dependencies and process
Even though some packages are 'safer' to update,
we cannot rely on having the CI discovering it on
random integrations. Pinning the remaining packages
and ordering them a bit.

The update of the 'build' package (to create wheels)
required the replacement of 'build.pep517' by
the new 'pyproject_hooks' module, and other
modification to the wheel artifacts.

The removed dependencies are currently placed at the tool level
requirements.txt so it's not like they are not needed anymore.

Test and CI scripts were adapted in order to rely on the 'dist'
directory rather than the 'dist_new' one, removing the old
step of creating the wheels with 'setup.py bdist_wheel'.

The entry points (console scripts) that we used to have
in the 'setup.py' were moved to the 'pyproject.toml' in order
to advance towards not relying on the setup.py file.

Flake8 issues were addressed in the different files
that this patch modified.

Change-Id: I83480c1920206e11fcb9a45264b6beaf6f8b686b
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit db554a0cb574ddbb2bb27de7649976bcaa85cdfe)