build: check module consistency for wheels
A few modules have been excluded in releases because we haven't
noticed they are missing from the wheels. This adds
a small check to the wheel creation, so we can be aware of
this situation.

A similar check is done with the README content, which is
what we list we are including in the wheels.

Pick-to: 6.6
Change-Id: I719e7d35b1466329a537c6095af01e705b7c6bea
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>