build: check module consistency for wheels
A few modules have been excluded in releases because we haven't
noticed they are missing from the wheels. This adds
a small check to the wheel creation, so we can be aware of
this situation.

A similar check is done with the README content, which is
what we list we are including in the wheels.

Change-Id: I719e7d35b1466329a537c6095af01e705b7c6bea
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit 39ab1a9f93dba84af6e179edb9dfb7c3a1ca3cf4)