LazyInit: Make imports computable from offset constants
The function cpythonTypeNameExt uses getTypeIndexVariableName
to compute the offset to a class variable.

For LazyInit we want to change this, creating an expression
that can compute the referenced types on demand when no
value is found at the offset. These types are hidden in the offset
expression.

By no longer using "toUpper" here, we can easily compute the
needed imports from the offset constant name.
We will keep both versions until PySide7.

[ChangeLog][PySide6] Type index constants are no longer in uppercase.
                     Uppercase is retained until deprecation in PySide7.

Task-number: PYSIDE-2404
Change-Id: I422ead0540a7bdcb4c7ac9905c8e6ab71b5b0634
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>