QTimer: Call C++ function for singleShot timers
The current implementation of singleshot timers is very old and bypasses
the C++ function. Instead, a timer object is manually created and
started. This incurs a performance penalty, as this bypasses the
optimized code path for 0 timers that eschews a timer object in favor of
directly calling QMetaObject::invokeMethod. This is now fixed, and for 0
timers, the C++ function is called directly.

Change-Id: Idfed06d60eb34355242818ac2df46f75dd27353c
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>