QObject: Add connect() overload with context arg
On C++, there is an overload of QObject::connect() that allows passing a
context object. This creates a connection in the context object's event
loop instead of that of the caller. This implements said overload in
Python.

Change-Id: Ia6a098a04470741efd450818c31aa5bffb1bd9f8
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
(cherry picked from commit acab25a3ccb836818e5089b23d40196bc7414b7a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>