QTimer: Fix singleShot overload with context
The singleShot overload with context had a problem where if the functor
was a slot of a QObject, it would not have a sender when it should. To
fix this, the newly added QObject.connect() overload with context is
used.

Change-Id: I654c09efb0d1b37ea0c014e9f17cd5e1913d1a96
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
(cherry picked from commit 6b227000f28c82e189f3cc26275ca7a930f75e2f)
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>