Long live PyObjectHolder
Add a class PyObjectHolder for passing around Python objects in C++.
It is mainly intended for callables. There are some cases in which
callables are invoked several times or not at all.

The usual construct of passing a lambda which decrefs the callable
after the invocation only works in the case when there is exactly one
single shot invocation. Otherwise, the callable leaks if there is no
invocation or a crash occurs on multiple invocation due to repeated
decref.

To fix this, introduce a class PyObjectHolder which holds a PyObject
with a reference and implements move/copy semantics as well as
destruction with the GIL held. It serves as a base class for functors
which implement a call operator for the invocation.

For the singleshot case, release() can be called after the invocation
as not to unnecessarily delay the release.

Port over the known cases where multiple invocation is documented to
occur to use it.

Change-Id: I2f4c7d497d4be8e48946f77a5a40dc9712dc05dd
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>