shiboken6: Extract common C++ function attributes
Extract attributes common to C++ functions from
AbstractMetaFunction's attributes to a flags type in
codemodel_enums.h for re-use in AbstractMetaFunction,
FunctionModelItem and AddedFunction.

A lot of boolean setters and getters can then be removed and
the flags easily transferred from code model to the meta
language code.

Task-number: PYSIDE-2602
Pick-to: 6.7
Change-Id: I9da1bb0d70051cd6bb3113290e8ccedfe4d13908
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>