Fix connect() for QObject callables
The addition of the QObject.connect() overload that takes a context
resulted in an automatic overload ordering that prevented the right
overload from ever being called if the callable was a QObject. Set a
manual order to fix this.

Fixes: PYSIDE-2627
Change-Id: I863f6caaed926b48bc412a29541c1d431ebd50b0
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
(cherry picked from commit e8b8ca6846530ceabcfb713aaa2804da18711eda)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>