Lazy Import: Ensure type creation functions being idempotent
Under circumstances it was possible to create a type
twice. There would be many more changes necessary to
ensure that this can never happen.

Very simple to prevent this was by checking that
a second call of the creation function does nothing,
IOW making the functions idempotent.

Tested with and without laziness.

Task-number: PYSIDE-2404
Change-Id: I70e3335a12f0cbe1111febaace7275e87c843bd4
Fixes: PYSIDE-2674
Pick-to: 6.7
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>