shiboken6: Add a placeholder for the base class to polymorphic-id-expression
Add %B for base class in addition to %1 for the class itself, which
is not useful and may lead to undefined behavior.

As a drive-by fix up the hitherto unused "polymorphic-base" which
is a boolean indicating the base class.

Task-number: PYSIDE-2675
Change-Id: I078191dc7b4c686b196fe58d6df9a249cdf2b151
Reviewed-by: Christian Tismer <tismer@stackless.com>
(cherry picked from commit 4a04afc95d348c8f26919a2f084ee80fa404e7d9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>