shiboken6: Remove function attribute FinalInTargetLang
The attribute does not have any impact on the generated code; it merely
causes some warnings to be printed when fixing the class functions
for rare cases. AbstractMetaFunction::isModifiedFinal() determines
the generation of virtual functions instead.

The function query flags FinalInTargetLangFunctions and
VirtualInTargetLangFunctions can be removed as well. They were
both used in AbstractMetaClass::functionsInTargetLang() which
is used by the documentation generator only.

Pick-to: 6.7
Task-number: PYSIDE-2602
Change-Id: Ib0f0de41008b655a84637ab28f1046aa3e098b76
Reviewed-by: Christian Tismer <tismer@stackless.com>