shibokenmodule: Expose dumpTypeGraph(), dumpWrapperMap() from the binding manager
This is helpful for gaining some insight into the instantiated
wrappers and registered types.

Move the existing graph generating code out of the graph class
and format the nodes with a short name and tooltip and expose it
as dumpTypeGraph().

Similarly, rewrite expose showWrapperMap() to use streams and
expose it as dumpWrapperMap().

Change-Id: I5cff442b7285388403dcf1c9f96fa8808c7b2d05
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
(cherry picked from commit 193769216f60f87feb20bbffa832cc159bbe525c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>