shiboken6: Lazy-initialize the dependency graph
Change the dependency graph in the BindingManager from
PyTypeObject *-based nodes to nodes which are wrapping the
TypeInitStruct * and hashing on the type name. This allows for
creating the types on demand when walking along the edges and calling
the type discovery functions.

This only creates the required types instead of the entire
lazy group of polymorphic classes.

The graph is now populated by from the generated code using a
function named initInheritance() instead of dynamically
from introduceWrapperType.

Task-number: PYSIDE-2404
Task-number: PYSIDE-2675
Pick-to: 6.7
Change-Id: I030d4957c221f4defbb7cc52a6927287b70d9864
Reviewed-by: Christian Tismer <tismer@stackless.com>