PEP 697: Use the new type extension provision, final
The new embedded extra fields were not correctly
addressed when a subclass of the meta type was used.

This was fixed by using SbkObjectType explicitly instead
of the type of an object which might be derived. This
also explains the complexity of the Python API.

The famous bug_825 is now fixed, too.

Checked with Valgrind on Ubuntu.

Task-number: PYSIDE-2230
Change-Id: Icf99ed6faab226c72bdd09f4c69018d958ede85e
Fixes: PYSIDE-2676
Pick-to: 6.6
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit b782c3054fe28f6f23ecab064a700f1da917d88f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>