Fix class hierarchies not working with legacy qmlRegisterType()
In order to work with the new QML registration code, the legacy
qmlRegisterType() function set QMetaClassInfo items on the meta object
of the QObject type to be registered. This caused the meta object to
be recreated in PySide's dynamic meta object builder, breaking the
class inheritance information.

To fix this, use a separate dummy meta object to provide
the information.

Amends 91bf9aa10faad14de557136664f58005c935d11c.

Pick-to: 6.7
Fixes: PYSIDE-2709
Change-Id: Icbd56759279bc8fcf89705af013db560babe4065
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>