Improve error handling when calling Python overrides of virtual methods
When calling into Python overrides of virtual methods,
there is mostly no Python context and so, potential errors,
for example caused by misguided duck-punching, appear as
error messages without context. To fix this, add an
optional context string parameter to storeErrorOrPrint()
and a special method for virtuals that formats class and
method name.

Change-Id: Ibbaf63c163edeb9bf8fcd364c108179c2ac6a069
Reviewed-by: Christian Tismer <tismer@stackless.com>