Remove the usage of Py_SetProgramName
Py_SetProgramName is deprecated since 3.11 and there is no planned
direct replacement. There is a recommendation of using PyConfig,
which implies a more verbose approach, so we drop the lines
instead of replacing it.

This:

    Py_SetProgramName(L"module-test");

can be replaced by:

    PyStatus status;
    PyConfig config;
    PyConfig_InitPythonConfig(&config);
    status = PyConfig_SetString(&config, &config.program_name, program_name);
    if (PyStatus_Exception(status)) {
        goto exception;
    }
    status = Py_InitializeFromConfig(&config);
    if (PyStatus_Exception(status)) {
        goto exception;
    }
    exception:
        PyConfig_Clear(&config);
        Py_ExitStatusException(status);

Note: PyConfig's approach is not part of the Stable ABI,
meaning it will not be possible to use it once Py_SetProgramName
is removed in 3.13

Change-Id: I0734d502469ae4f28f4c211bdfbfdbeae82b1375
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>