shiboken6: Fix type names of smart pointers in namespaces in the cppApi array
Set the correct parent entry on the smart pointer instantiation type entry.

This allows for removing a special case in the source generator and
namespaces will be correctly reflected. It is also the prerequisite
for a follow-up change making it possible to use global namespaces for
smart pointers in namespaces.

Also remove the setting of the enclosing class and the associated flag
InheritTemplateFlag::SetEnclosingClass from
AbstractMetaBuilderPrivate::inheritTemplate(). Instead, always set the
enclosing class in ApiExtractorPrivate::addInstantiatedSmartPointer()
instead, depending on the type of the parent entry.

Task-number: PYSIDE-2404
Task-number: PYSIDE-2768
Pick-to: 6.7
Change-Id: I921e255c2ceaeb9eaf84e5ac1c4953bcf7d09b57
Reviewed-by: Christian Tismer <tismer@stackless.com>