Desktop Deployment: Fix error
In the scenario where 'pyside6-deploy --init' is used on the first run
and for the consequent invocations, only running just 'pyside6-deploy'
without any additional options, the deployment should have worked.

This currently had a bug where it overrides the main Python entrypoint
file with the default option of 'main.py'. This issue is caused due to
the order of 'if' loop in the changed function.

Pick-to: 6.7
Task-number: PYSIDE-1612
Change-Id: I605c4b9ff2035e85c0b5f73049a3ecc84d0fd80c
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>