Shiboken: Get rid of the fullName field in generated code
This patch does the actual removal of the fullName field.
Unfortunately, not all fields can be removed but only
the ones which are reachable by the TypeInitStructs.
(meanwhile less than 500)

This result is pretty much better (at least on macOS),
so we would not want to optimize it any further.

Many thanks to Friedemann's help with improving the
generator and removing the quirks that I could not solve.

Task-number: PYSIDE-2701
Change-Id: Ie350766c450da1d86359c4e24c313ade3146233f
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>