QtAsyncio: Don't cancel handle
It is not necessary to cancel a task's handle when cancelling said task,
as the step function already handles the cancellation. Furthermore,
cancelling the handle can cause issues with TaskGroups if an exception
is raised inside the TaskGroup's context, where the tasks inside the
group are not allowed to properly cancel and thus block the program.

Pick-to: 6.7
Task-number: PYSIDE-769
Fixes: PYSIDE-2799
Change-Id: I1ab782380ffbe8c1ec40aba22b9ed2ebdcb1498f
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>