shiboken6: Refactor GeneratorContext usage
Pass GeneratorContext to more functions which will be required in a
follow-up change.

Add a GeneratorContext::Type enumeration value indicating it is a
default-constructed context used for global functions. Add some
asserts on the presence of a class and fix violations.

Add an overload QString CppGenerator::typeInitStruct(GeneratorContext)
that handles smart pointers.

Task-number: PYSIDE-2701
Pick-to: 6.7
Change-Id: I453d584afed11917c2a45e0440d3aa6235650e46
Reviewed-by: Christian Tismer <tismer@stackless.com>