pyi_generator: Remove duplicate annotations line
The future annotations line was erroneously added to the license text,
resulting in it being added twice to pyi files. We keep it in the
license text, remove the other code where it was being added, and
remove the USE_PEP563 variable, as it is now always true.

Task-number: PYSIDE-2786
Change-Id: Ic83b489f3ac81dd0730b87c5323a757cb5e0008c
Reviewed-by: Christian Tismer <tismer@stackless.com>