example_gallery: Handle examples which have the pyproject file in the "doc" dir
Examples like samplebinding, etc. have a CMakeLists.txt and only
a dummy pyproject file in the "doc" dir, which caused the zipping to
fail.

Modify make_zip_archive() to take the complete target path as an
argument and remove the move command.

Change one level up if the pyproject file is in "doc".

Pick-to: 6.7
Change-Id: Ibf6064e6e0e6a4e24e25348d3a4b52d361ae9560
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>