doc: improve a bit the gettingstarted page
Add explanation of Qt Widgets and Qt Quick,
remove outdated information, and move the notes
around where they make more sense.

Add a banner for both technologies, and adding
a screenshot of the Quick example.

At the end, add a section to point the readers
to the Examples and Tutorials sections.

Change-Id: Ie3b1856b7150a83633f9747a22057d902fb44b48
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>