type hints: Evaluate multiple signatures and Union Expressions
The multiple signatures suffer from the problem that
many are reported as error, because one larger type
is evaluated before another type.

An Example: PySide6.QtCore.QCborSimpleType is shadowed by int
when int is listed first. That is due to the mro:

  int.mro()              [<class 'int'>, <class 'object'>]

  QCborSimpleType.mro()  [<enum 'QCborSimpleType'>, <enum 'IntEnum'>,
                          <class 'int'>, <enum 'ReprEnum'>,
                          <enum 'Enum'>, <class 'object'>]

We are implementing a topological sort by mro length that
puts the list into the right order.

Result is from 226 down to 30.

Change-Id: I155e2169c3c8ab4afd5b7bb8f3411ccf6467d440
Task-number: PYSIDE-2846
Fixes: PYSIDE-2687
Fixes: PYSIDE-2886
Pick-to: 6.8
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>