type hints: Install a permanent test for error-free pyi files
Testing Mypy on the whole project costs some time,
about 33 s on a Mac Studio Max M1 in debug mode.
On CI with release Windows, this is about 8 sec.

An option "--skip-mypy-test" is added to save time.
The mypy tool is only enforced to exist in COIN.
Otherwise it is optional.

Task-number: PYSIDE-2846
Change-Id: Ibbabbf6b893537ac923fbbaba9ff832ebd675a21
Fixes: PYSIDE-2672
Fixes: PYSIDE-2886
Pick-to: 6.8
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>