type hints: Simplify the NoneType / None issue
Under some circumstances, a type 'NoneType' is generated when
calling eval() on a string. This is correct, but is generally
to be simplified to 'None'.

This is extracted from the type hint chain as it is not
related to mypy errors.

Task-number: PYSIDE-2846
Change-Id: Ib705b32a82c46c597c2724092a38a1861319ca16
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>
(cherry picked from commit 9af99c2fd8ed6b8b8959b1b33717989d51555e3d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>