QtAsyncio: Improve logging when an Exception occurs in a task
When an exception occurs in a task, the logging is not helpful since it prints out nothing specific about the error that occurred. Catching the traceback and printing it out with the task in which it happened brings a lot of details to resolve the exception faster.

Task-number: PYSIDE-3001
Change-Id: If199da28a37406779ac5bec178fe756d1757b08c
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit c253dcda92cc2efb45197190beaccb926afb120c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>