shiboken6: Fix passing QFlags by const-ref
The type resolution code searched the result of TypeInfo::toString()
in the type database, which may contain qualifications like
"const Foo &".

To fix this, add a function returning the name with instantiations
only and use that. As a drive-by, add a convenvience function
returning the qualified name as a string and simplify the code
accordingly.

Change-Id: I071e8dc3f65257e300007d53688c442a683813e1
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>
(cherry picked from commit 779e6c32b1580a658b7906accb391adb799c6fe6)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>