shiboken6/DocGenerator: Improve handling of missing detailed descriptions
Refactor the functions to extract brief/detailed texts of the
qdoc parser to return a pair of strings and do extended checking
for space and empty elements. Add conditions checking on it to
the doc generator.

Adapt the test which had characters outside elements in the description.

Pick-to: 6.9 6.8
Task-number: PYSIDE-1106
Change-Id: I9447270ba4280d0e6e42c3ebe9f76c332581dd60
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
(cherry picked from commit 21992dafe7b126ee65a7e775cbdac94908184d29)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>